+++ src/expat_fatcat/expat_fatcat.py
@@ -9,13 +9,18 @@ import quandl
     
 class FatcatCalculator():
     
-    def __init__(self, FXRateConverter):
-        self.fx_rate_converter = FXRateConverter('USD')
+    def __init__(self, fx_rate_converter):
+        self.fx_rate_converter = fx_rate_converter
+        self._check_converter()
+        
+    def _check_converter(self):
+        if self.fx_rate_converter.to_currency != 'USD':
+            raise ValueError('invalid "to_currency": Fatcat only converts to USD')
         
     def _parse_date(self, date_string, date_format='%Y-%m-%d'):
         return datetime.strptime(date_string, date_format)
     
-    def _get_rate(self, from_currency, date_string, date_format='%Y-%m-%d'):
+    def get_rate(self, from_currency, date_string, date_format='%Y-%m-%d'):
         date = self._parse_date(date_string, date_format)
         rate = self.fx_rate_converter.get_rate(from_currency, date)
         return rate
@@ -35,13 +40,34 @@ class AbsRateConverter(ABC):
         pass
 
 
-    
+class DummyRateConverter(AbsRateConverter):
+
+        def __init__(self, to_currency):
+            self.to_currency = to_currency
+
+
+        def _get_call_str(self, from_currency):
+            if from_currency == 'FOO':
+                return 'ECB/FOO' + self.to_currency
+            else:
+                raise NotImplementedError('Only FOO to {} rates are implemented'.format(self.to_currency))
+        
+
+        def _dummy_api_call(self, call_str):
+            return 1.125
+
+
+        def get_rate(self, from_currency, date):
+            '''Returns dummy exchange rate'''
+            call_str = self._get_call_str(from_currency)
+            return self._dummy_api_call(call_str)
+        
+
 class QuandlUSDRateConverter(AbsRateConverter):
 
     def __init__(self):
         self.to_currency = 'USD'
         self._set_key()
-        self._set_call()
     
     def _set_key(self):
         if 'QUANDL_KEY' in os.environ:
